requirements.txt
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.36.0
diffusers>=0.25.0
peft>=0.7.0
safetensors>=0.4.0
Pillow>=10.0.0
huggingface-hub>=0.19.0
flash-attn>=2.3.0